💄 design: (FE) - chat 입력창 구분선 추가 및 스타일 파일 분리
