+++ frontend/src/components/@drawer/ChatDrawer/ChatDrawer.tsx
@@ -6,8 +6,14 @@ import { useRecoilState, useRecoilValue } from 'recoil';
 import { meInRoomState } from '@store/user.store';
 import Button from '@components/@shared/Button/Button';
 import { chatListState } from '@store/chatList.store';
-import { css } from '@emotion/react';
-import { flexColumn, flexRow } from '@styles/globalStyle';
+import {
+	chatContentStyle,
+	chatDrawerStyle,
+	chatInputStyle,
+	chatItemStyle,
+	chatListContainer,
+	chatNicknameStyle,
+} from './ChatDrawer.style';
 
 export enum ChatTarget {
 	EVERYONE = 'everyone',
@@ -58,63 +64,12 @@ const ChatDrawer = () => {
 					);
 				})}
 			</div>
-			<div>
-				<form css={chatInputStyle} onSubmit={sendChat}>
-					<TextField value={chatContent} onChange={handleChangeChatContent}></TextField>
-					<Button onClick={sendChat}>Send</Button>
-				</form>
-			</div>
+			<form css={chatInputStyle} onSubmit={sendChat}>
+				<TextField value={chatContent} onChange={handleChangeChatContent}></TextField>
+				<Button onClick={sendChat}>Send</Button>
+			</form>
 		</div>
 	);
 };
 
 export default ChatDrawer;
-
-const chatDrawerStyle = (theme) => css`
-	${flexColumn({ justifyContent: 'space-between' })};
-	width: 100%;
-	height: calc(100% - 56px);
-`;
-
-export const chatListContainer = (theme) => css`
-	display: flex;
-	flex-direction: column-reverse;
-	gap: 16px;
-
-	width: 100%;
-	height: 100%;
-	padding: 12px 0px;
-
-	overflow: auto;
-`;
-
-export const chatItemStyle = (theme, isMe) => css`
-	${flexColumn({ gap: '4px', alignItems: isMe ? 'end' : 'flex-start' })};
-
-	width: 100%;
-	text-align: ${isMe ? 'right' : 'left'};
-`;
-
-export const chatNicknameStyle = (theme) => css`
-	color: ${theme.colors.white};
-	font-size: ${theme.fontSize.xSmall};
-`;
-
-export const chatContentStyle = (theme, isMe) => css`
-	width: 80%;
-	padding: 12px;
-
-	color: ${isMe ? theme.colors.white : theme.colors.black};
-	background-color: ${isMe ? theme.colors.primary : theme.colors.gray3};
-
-	font-size: ${theme.fontSize.small};
-	border-radius: ${theme.borderRadius};
-`;
-
-export const chatInputStyle = (theme) => css`
-	${flexRow({ gap: '8px', alignItems: 'center' })};
-
-	width: 100%;
-	height: 56px;
-	padding-top: 16px;
-`;
